submission_worker/requirements.txt
pika==1.2.0
requests>=2.30.0,<3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
responses==0.24.1
pytest-cov==4.1.0 